binning modification: dict keys
I have only changed the keys of the dictionary for the bins. Before the key of each bin was the center of the bin. Now the bin dict keys are just integers which makes everything much cleaner and easier to handle later (floating point precision problems can arise if not integer).

Other than that no on the changes were implemented.